+++ trixnity-messenger/src/androidMain/kotlin/de/connect2x/trixnity/messenger/util/FileDescriptor.android.kt
@@ -7,17 +7,19 @@ import de.connect2x.lognity.api.logger.Logger
 import de.connect2x.trixnity.messenger.i18n.I18n
 import de.connect2x.trixnity.utils.ByteArrayFlow
 import de.connect2x.trixnity.utils.byteArrayFlowFromSource
-import io.ktor.http.*
+import io.ktor.http.ContentType
+import io.ktor.http.fromFilePath
 import okio.Buffer
 import okio.source
 
 private val log: Logger = Logger("de.connect2x.trixnity.messenger.util.UriFileDescriptorKt")
 
-class UriFileDescriptor(private val context: Context, internal val fileUri: Uri, i18n: I18n) : FileDescriptor {
+class UriFileDescriptor(private val context: Context, private val fileUri: Uri, i18n: I18n) : FileBackedFileDescriptor {
 
     private val computedFileInfo = getComputeFileInfo(fileUri)
 
     override val fileName: String = computedFileInfo?.fileName ?: i18n.commonUnknown()
+    override val filePath: FilePath = FilePath(fileUri)
     override val fileSize: Long? = computedFileInfo?.fileSize
     override val mimeType: ContentType? =
         computedFileInfo?.mimeType?.let(ContentType.Companion::parse)
@@ -43,5 +45,7 @@ class UriFileDescriptor(private val context: Context, internal val fileUri: Uri,
             }
             .getOrNull()
 }
+
+actual class FilePath(val uri: Uri)
 
 data class ComputedFileInfo(val fileName: String, val fileSize: Long?, val mimeType: String?)