+++ trixnity-messenger/src/iosMain/kotlin/de/connect2x/trixnity/messenger/util/FileDescriptor.ios.kt
@@ -2,6 +2,7 @@ package de.connect2x.trixnity.messenger.util
 
 import de.connect2x.trixnity.utils.BYTE_ARRAY_FLOW_CHUNK_SIZE
 import de.connect2x.trixnity.utils.ByteArrayFlow
+import de.connect2x.trixnity.utils.toByteArrayFlow
 import io.github.vinceglb.filekit.PlatformFile
 import io.github.vinceglb.filekit.path
 import io.github.vinceglb.filekit.size
@@ -12,11 +13,22 @@ import io.ktor.http.fromFilePath
 import kotlinx.coroutines.flow.flow
 import kotlinx.io.InternalIoApi
 import kotlinx.io.buffered
+import kotlinx.io.files.Path
 import kotlinx.io.readByteArray
+import platform.Foundation.NSURL
 
-class FileKitFileDescriptor(internal val file: PlatformFile) : FileDescriptor {
+class ByteArrayFileDescriptor(private val path: Path, private val data: ByteArray) : FileDescriptor {
+    override val fileName: String = path.name
+    override val fileSize: Long? = data.size.toLong()
+    override val mimeType: ContentType? = ContentType.fromFilePath(fileName).firstOrNull()
+    override val content: ByteArrayFlow = data.toByteArrayFlow()
+}
+
+class FileKitFileDescriptor(internal val file: PlatformFile) : FileBackedFileDescriptor {
+
     override val fileName: String = file.path.toPath().name
-    override val fileSize: Long = file.size()
+    override val filePath: FilePath = FilePath(file.nsUrl)
+    override val fileSize: Long? = file.size()
     override val mimeType: ContentType? = ContentType.fromFilePath(fileName).firstOrNull()
 
     @OptIn(InternalIoApi::class)
@@ -28,3 +40,5 @@ class FileKitFileDescriptor(internal val file: PlatformFile) : FileDescriptor {
         }
     }
 }
+
+actual class FilePath(val nsUri: NSURL)