+++ trixnity-messenger/src/jvmMain/kotlin/de/connect2x/trixnity/messenger/util/FileDescriptor.jvm.kt
@@ -6,10 +6,13 @@ import io.ktor.http.fromFilePath
 import okio.FileSystem
 import okio.Path
 
-class PathFileDescriptor(val path: Path, private val fileSystem: FileSystem) : FileDescriptor {
+class PathFileDescriptor(private val path: Path, private val fileSystem: FileSystem) : FileBackedFileDescriptor {
 
     override val fileName = path.name
+    override val filePath = FilePath(path)
     override val fileSize = fileSystem.metadataOrNull(path)?.size
     override val mimeType = ContentType.fromFilePath(fileName).firstOrNull()
     override val content = byteArrayFlowFromSource { fileSystem.source(path) }
 }
+
+actual class FilePath(val path: Path)